added plot function to core.py
